+++ xlators/transport/tcp/src/tcp.c
@@ -849,7 +849,7 @@ brick_open (struct xlator *xl,
     struct file_context *brick_ctx = calloc (1, sizeof (struct file_context));
     brick_ctx->volume = xl;
     brick_ctx->next = NULL;
-    *(int *)&brick_ctx->context = data_to_int (dict_get (&reply, "FD"));
+    *(long *)&brick_ctx->context = data_to_int (dict_get (&reply, "FD"));
     
     while (trav->next)
       trav = trav->next;
